+++ tests/examples/test_robot_arm.py
@@ -1,0 +1,7 @@
+# Copyright 2026 MIT Lincoln Laboratory
+# SPDX-License-Identifier: MIT
+
+from pydgens.examples.robot_arm import main
+
+def test_robot_arm_smoketest():
+    main()